+++ gradle/libraries.toml
@@ -38,7 +38,7 @@ ktor = "2.3.7"
 moshix = "0.14.1"
 node-fetch = "2.6.7"
 node-ws = "8.14.2"
-okio = "3.6.0"
+okio = "3.9.0"
 okhttp = "4.11.0"
 rx-android = "2.0.1"
 rx-java2 = "2.2.21"
