+++ libraries/apollo-ast/src/concurrentMain/kotlin/com/apollographql/apollo3/ast/host.concurrent.kt
@@ -1,6 +1,7 @@
 package com.apollographql.apollo3.ast
 
 import okio.FileSystem
+import okio.SYSTEM
 
 internal actual val HOST_FILESYSTEM: FileSystem
   get() = FileSystem.SYSTEM
