+++ libraries/apollo-ast/src/filesystemMain/kotlin/com/apollographql.apollo3.ast/introspection/introspection_reader.filesystem.kt
@@ -1,0 +1,19 @@
+@file:JvmMultifileClass
+@file:JvmName("Introspection")
+
+package com.apollographql.apollo3.ast.introspection
+
+import com.apollographql.apollo3.annotations.ApolloExperimental
+import com.apollographql.apollo3.ast.HOST_FILESYSTEM
+import okio.Path
+import okio.buffer
+import kotlin.jvm.JvmMultifileClass
+import kotlin.jvm.JvmName
+
+@ApolloExperimental
+fun Path.toIntrospectionSchema(): IntrospectionSchema {
+  return HOST_FILESYSTEM
+      .source(this)
+      .buffer()
+      .toIntrospectionSchema(toString())
+}
