+++ libraries/apollo-ast/src/filesystemTest/kotlin/test/validation/inputTypeDeprecatedField.graphqls
