--- libraries/apollo-ast/src/linuxMain/kotlin/com/apollographql/apollo3/ast/host.linux.kt
@@ -1,6 +1,0 @@
-package com.apollographql.apollo3.ast
-
-import okio.FileSystem
-
-internal actual val HOST_FILESYSTEM: FileSystem
-  get() = FileSystem.SYSTEM
