+++ libraries/apollo-ast/src/wasmJsTest/kotlin/com/apollographql/apollo3/graphql/ast/test/cwd.wasmJs.kt
@@ -1,0 +1,4 @@
+package com.apollographql.apollo3.graphql.ast.test
+
+internal actual val CWD: String
+  get() = TODO("Not yet implemented")
